requirements.txt
pymongo
customtkinter
python-dotenv